+++ templates/index.php
@@ -6,3 +6,7 @@ use OCP\Util;
 $appId = Application::APP_ID();
 
 Util::addScript($appId, $appId . '-main');
+?>
+
+<div id="app-content-vue"></div>
+<div id="app-navigation-vue"></div>